requirements.txt
absl-py==2.0.0
altair==5.1.2
altgraph==0.17.4
annotated-types==0.5.0
anyio==4.0.0
Appium-Python-Client==3.1.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
autopep8==2.0.4
awscli==1.29.56
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
blis==0.7.11
botocore==1.31.56
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
confection==0.1.3
constantly==23.10.4
cryptography==41.0.5
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.8
Cython==3.0.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.5
docutils==0.20.1
einops==0.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.4
flake8==6.1.0
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
future==0.18.3
gast==0.5.4
gensim==4.3.2
google-auth==2.23.1
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.19.0
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keras==2.15.0
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
line-profiler==4.1.1
llvmlite==0.41.0
lxml==4.9.1
macholib==1.16.2
Markdown==3.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
ml-dtypes==0.2.0
MouseInfo==0.1.3
mpmath==1.2.1
murmurhash==1.0.9
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==3.1
nodejs==0.1.1
notebook==6.4.12
notebook_shim==0.2.3
npm==0.1.1
numba==0.56.2
numpy==1.26.0
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
optional-django==0.1.0
outcome==1.3.0.post0
overrides==7.4.0
packaging==21.3
pandas==2.1.3
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
pathy==0.10.2
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.0.0
pluggy==1.4.0
preshed==3.0.8
pretty-errors==1.2.25
prince==0.12.1
prometheus-client==0.14.1
prompt-toolkit==3.0.31
Protego==0.3.0
protobuf==4.23.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyDispatcher==2.0.7
pyelftools==0.29
pyflakes==3.1.0
PyGetWindow==0.0.9
Pygments==2.13.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyOpenSSL==23.3.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.30
pyserial==3.5
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
python-json-logger==2.0.7
pytweening==1.0.7
pytz==2022.2.1
PyYAML==5.4.1
pyzmq==24.0.0
queuelib==1.6.2
redis==5.0.1
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.7.2
rubicon-objc==0.4.7
s3transfer==0.7.0
scikit-learn==1.1.2
scipy==1.11.3
Scrapy==2.11.0
seaborn==0.12.0
selenium==4.15.2
Send2Trash==1.8.2
service-identity==23.1.0
six==1.16.0
sklearn==0.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.7
stack-data==0.5.0
sympy==1.11.1
tensorboard==2.15.1
tensorboard-data-server==0.7.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.3.0
terminado==0.15.0
thinc==8.1.12
threadpoolctl==3.1.0
tinycss2==1.1.1
tldextract==5.0.1
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.66.1
traitlets==5.10.1
trio==0.23.1
trio-websocket==0.11.1
Twisted==22.10.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.5
virtualenv==20.19.0
w3lib==2.1.2
wasabi==1.1.2
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yapf==0.32.0
zope.interface==6.1
